modloader updates, fun stuff and weapon adjuster things

+++ source/code/plugin/eCustomTableOfContents.cpp
@@ -10,12 +10,10 @@ char& unk_67D000 = *(char*)0x67D000;
 
 void eCustomTableOfContents::InitHooks()
 {
-	//InjectHook(0x4D5090, LoadFile, PATCH_JUMP);
 	InjectHook(0x4D61B0, GetTOCFileSize, PATCH_JUMP);
 }
 
 int eCustomTableOfContents::GetTOCFileSize(char * name)
 {
-	//printf("File %s\n", name);
-	return std::filesystem::file_size(name);
+	return (int)std::filesystem::file_size(name);
 }
